LevitechDemo/requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.20

# Document processing
pymupdf>=1.23.0
python-docx>=1.1.0
ocrmypdf>=16.0.0

# Vector search
chromadb>=0.4.0
openai>=1.3.0

# Hybrid search
rank-bm25>=0.2.2

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0
